feat: Graph파일코드 합치기&Tap기능으로 각 그래프 확인
